backend/requirements.txt
Django>=5.1,<5.3
djangorestframework>=3.15,<3.16
django-cors-headers>=4.5,<4.7
djangorestframework-simplejwt>=5.3,<5.5
